requirements.txt
# Django
django==4.0.2
djangorestframework==3.13.1
django-extensions==3.1.5
django-storages==1.10.1
django-cors-headers==3.11.0
django-filter==21.1
django-redis==5.2.0
djangorestframework-simplejwt==5.0.0
channels==3.0.4

# Serving
gunicorn==20.1.0
uvicorn[standard]==0.17.0

# Integrations
boto3==1.20.41
sentry-sdk==1.5.3
slack_sdk==3.13.0
ddtrace==0.58.5
datadog==0.43.0
twilio==7.9.3

# Database
psycopg2==2.9.3

# Code clean
flake8==4.0.1
autopep8==1.6.0

# Application
requests==2.27.1
python-decouple==3.5
pdfrw==0.4.0
celery==5.2.3
Jinja2==3.0.3

# Redis
redis==4.1.1
walrus==0.9.0
channels-redis==3.2.0

# Pydantic
email-validator==1.1.3
pydantic==1.9.0
phonenumbers==8.12.50
iso3166==2.1.1

# Testing
factory-boy==3.2.1
pytest==6.2.5
pytest-django==4.5.2
pytest-mock==3.7.0
freezegun==1.1.0

# Logging
python-json-logger==2.0.2
colorlog==6.6.0
django-log-request-id==2.0.0

# Debugging
ipdb==0.13.9
ipython==8.0.1

# Dependent depdencies
cryptography==3.2.1

# Infra ops
fire==0.4.0
inquirer==2.7.0
paramiko==2.9.2
awscli==1.22.41
pynpm==0.1.2
